Eye fixation detection
When the eye gaze out of the fixation rect for 2 sec the whole block was abandoned. At the end of the block add one new block for abandoned block.
The problem is, when the eye gaze out of fixation for more than 1 time in one block, like 3 times, it will add 3 new blocks!!!!