CB stimulus perceived judgement--red bar
change the stimulus for patient 8, the judgement of perceived location using red bar---compatible with previous patient experiment.  The only different is after black disk fixation there is 1 sec blank for patient to saccade to perceived location and then the red bar appeared, the patient using red bar to present the perceived location.